requirements.txt
discord_py_interactions>=5.13.2
python-dotenv>=1.0.1
